moved files around and changed interactor so dependency rule isn't violated
