+++ src/data_access/InMemoryDataAccessObject.java
@@ -1,6 +1,7 @@
 package data_access;
 
 import entity.Property;
+import use_case.home.HomeSearchDataAccessInterface;
 
 import java.util.Map;
 import java.util.HashMap;
