+++ src/data_access/PropertyDataAccessObject.java
@@ -10,6 +10,7 @@ import okhttp3.Response;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+import use_case.home.HomeSearchDataAccessInterface;
 
 import java.io.*;
 import java.util.*;
